State: Move isJetpackSiteInDevelopmentMode selector to global selectors directory
